Refactor SuggestionService
