+++ src/main/java/petmatch/service/SuggestionService.java
@@ -1,0 +1,9 @@
+package petmatch.service;
+
+import petmatch.model.Profile;
+
+import java.util.List;
+
+public interface SuggestionService {
+    List<Profile> suggestProfiles(Profile myProfile);
+}
