+++ src/main/java/petmatch/service/implementation/SuggestionServiceImpl.java
@@ -1,4 +1,4 @@
-package petmatch.service.suggestion;
+package petmatch.service.implementation;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -8,6 +8,7 @@ import petmatch.model.Profile;
 import petmatch.service.InterestService;
 import petmatch.service.MatchService;
 import petmatch.service.ProfileService;
+import petmatch.service.SuggestionService;
 
 import java.time.LocalDate;
 import java.time.Period;
@@ -17,11 +18,12 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
-public class SuggestionService {
+public class SuggestionServiceImpl implements SuggestionService {
     private final ProfileService profileService;
     private final InterestService interestService;
     private final MatchService matchService;
 
+    @Override
     public List<Profile> suggestProfiles(Profile myProfile) {
         // Get the list of matched profiles
         List<Match> previousMatches = matchService.getPreviousMatches(myProfile);
